+++ app/config.js
@@ -25,9 +25,6 @@ module.exports = {
   // Force HTTP to redirect to HTTPS on production
   useHttps: 'true',
 
-  // Cookie warning - update link to service's cookie page.
-  cookieText: 'GOV.UK uses cookies to make the site simpler. <a href="#">Find out more about cookies</a>',
-
   // Enable or disable Browser Sync
   useBrowserSync: 'true'